+++ src/modules/texts/TextEdit/Context/index.ts
@@ -1,2 +1,1 @@
-export * as default from "./TextEdit.context";
-export * from "./TextEdit.context";
+export * from "./TextEditContext";
