+++ src/modules/texts/TextEdit/Context/types/index.ts
@@ -1,2 +1,1 @@
-export * as default from "./TextEdit.context.types";
-export * from "./TextEdit.context.types";
+export * from "./text-edit-context.types";
